requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.3
django-cors-headers==2.1.0
django-filter==1.0.4
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.6
Markdown==2.6.8
oauthlib==2.0.3
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
requests==2.18.4
urllib3==1.22
whitenoise==3.3.0
jsonschema==2.6.0
numpy==1.11.2
